separate appdata, sessionconfig, settings
