+++ src/main/remoteserverselectdialog/remoteserverselectdialog.ts
@@ -4,8 +4,8 @@
 import * as ejs from 'ejs';
 import { BrowserWindow, ipcMain } from 'electron';
 import * as path from 'path';
+import { appData } from '../config/appdata';
 import { ThemedWindow } from '../dialog/themedwindow';
-import { appData } from '../config/settings';
 
 export class RemoteServerSelectDialog {
   constructor(options: RemoteServerSelectDialog.IOptions) {
